% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_courses}
\alias{clean_courses}
\title{clean_courses
Clean the dataframe and save it in data/derived-data}
\usage{
clean_courses(courses)
}
\arguments{
\item{courses}{a data frame data/raw-data/CoursesVelomagg.csv}
}
\value{
courses
}
\description{
clean_courses
Clean the dataframe and save it in data/derived-data
}
