% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_data}
\alias{clean_data}
\title{CLean data}
\usage{
clean_data(x)
}
\arguments{
\item{data}{x l'objet que l'on veut clean}
}
\value{
x l'objet netooyé
}
\description{
fonction pour enlever les majuscules, la ponctuation, les espaces et les accents d'un objet
}
