% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_station}
\alias{clean_station}
\title{Clean, merge station dataframes and save in data/derived-data}
\usage{
clean_station(station, coord)
}
\arguments{
\item{station}{data frame data/raw-data/StationVelomagg.csv}

\item{coord}{data frame data/raw-data/DisponibiliteVelomagg.csv}
}
\value{
fusion data frame
}
\description{
Clean, merge station dataframes and save in data/derived-data
}
