% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Station_data.R
\name{Data_order}
\alias{Data_order}
\title{Reorganise l'ordre d'un data frame}
\usage{
Data_order(data, ascend, value)
}
\arguments{
\item{data}{data frame a ranger}

\item{ascend}{TRUE=ordre croissant, FALSE=ordre decroissant}

\item{value}{colone du data frame pour faire l'ordre}
}
\value{
data frame reorganisé
}
\description{
Reorganise l'ordre d'un data frame
}
