% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Station_data.R
\name{Data_station}
\alias{Data_station}
\title{Nombre de velos par station}
\usage{
Data_station(data, col, Fusion)
}
\arguments{
\item{data}{--> jeu de donnée}

\item{col}{--> colonne pour le compatge}

\item{Fusion}{}
}
\value{
data frame avec nombre de velo par station
}
\description{
Nombre de velos par station
}
