% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domaine_vital.R
\name{domaine_vitale}
\alias{domaine_vitale}
\title{Domaine vitale du cycliste}
\usage{
domaine_vitale(Course_velo2, Fusion)
}
\arguments{
\item{Course_velo2}{}

\item{Fusion}{}
}
\value{
ggplot avec le domaine vital
}
\description{
Domaine vitale du cycliste
}
