% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{download_course}
\alias{download_course}
\title{Download course
Download course csv
https://data.montpellier3m.fr/sites/default/files/ressources/TAM_MMM_CoursesVelomagg.csv}
\usage{
download_course()
}
\value{
No return value
}
\description{
Download course
Download course csv
https://data.montpellier3m.fr/sites/default/files/ressources/TAM_MMM_CoursesVelomagg.csv
}
