% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_data}
\alias{extract_data}
\title{Extract data
Extrait les donnees relatives à un mois}
\usage{
extract_data(courses, month)
}
\arguments{
\item{courses}{dataframe}

\item{month}{str exemple "06"}
}
\value{
dataframe
}
\description{
Extract data
Extrait les donnees relatives à un mois
}
