% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_map.R
\name{leaflet_map}
\alias{leaflet_map}
\title{Map leaflet}
\usage{
leaflet_map(Depstation, title)
}
\arguments{
\item{Depstation}{fichier station departure ou return}

\item{title}{titre graphique}
}
\value{
map leaflet
}
\description{
Map leaflet
}
