% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hist.R
\name{PLOT}
\alias{PLOT}
\title{Plot histograme Utilisation velo}
\usage{
PLOT(data, title)
}
\arguments{
\item{data}{Departure ou Return}

\item{title}{nom du graph}
}
\value{
ggplot histograme avec plotly
}
\description{
Plot histograme Utilisation velo
}
