% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_temps_temp_dist_month.R
\name{plot_temperature}
\alias{plot_temperature}
\title{Plot
Plot and save the plot}
\usage{
plot_temperature(courses_month, month)
}
\arguments{
\item{courses_month}{data}

\item{month}{}
}
\value{
No return value
}
\description{
Plot
Plot and save the plot
}
